Restrict,FindRow: initialize the row.length to 0 for when the response does not return anything + inverted the order of the fmid and table type arguments for consistency with the other backend methods

+++ mapiproxy/libmapistore/mapistore.h
@@ -174,7 +174,7 @@ int mapistore_modifyrecipients(struct mapistore_context *, uint32_t, uint64_t, s
 int mapistore_get_child_fids(struct mapistore_context *, uint32_t, uint64_t, uint64_t **, uint32_t *);
 int mapistore_deletemessage(struct mapistore_context *, uint32_t, uint64_t, uint8_t);
 int mapistore_get_folders_list(struct mapistore_context *, uint64_t, struct indexing_folders_list **);
-int mapistore_set_restrictions(struct mapistore_context *, uint32_t, uint8_t, uint64_t, struct mapi_SRestriction, uint8_t *);
+int mapistore_set_restrictions(struct mapistore_context *, uint32_t, uint64_t, uint8_t, struct mapi_SRestriction, uint8_t *);
 
 /* definitions from mapistore_processing.c */
 int mapistore_set_mapping_path(const char *);
